+++ src/Plugins/SmartStore.PayPal/changelog.md
@@ -1,5 +1,9 @@
 ﻿#Release Notes#
 
+##Paypal 2.2.0.1
+### New Features
+* Supports order list label for new incoming IPNs
+
 ##Paypal 1.22##
 ###Bugfixes###
 * PayPal Standard provider now using shipping rather than billing address if shipping is required